Refactor Home Page and Podcast Page
- Updated Home page to a new layout with a "Coming Soon" message and a call-to-action button.
- Added the logo to the navbar
